[NARS-27] BUG FIX: CalculateVolumeStatistic ont honoring correlation
The abstract statistics aggregator classes uses a correlation attribute to enable aggregations to be correlated by a user-specified attribute key. This subclass was not honoring that property. NOTE: this is a breaking change from the previous version because the property bucket interval has been removed.  This version of the NARs will prevent a clustered NiFi from starting if an existing version of this NAR exists anywhere on the canvas.