update dependencies
removed Java 11 incompatible method call 

removed `virtualMachineMetrics.fileDescriptorUsage()` because for this to work, a JVM argument is required: `--add-opens=jdk.management/com.sun.management.internal=ALL-UNNAMED`

while this is not a huge deal, the requirement to include the JVM argument is not something most deployments would be able to implement easily.  For that reason, the metric was removed.
added property descriptor to enable file descriptor metrics


update